CollisionSource defaultStringInterpolation improvements

+++ Sources/ErrorInfo/SharedTypes/TypeInfoOptions.swift
@@ -6,9 +6,9 @@
 //
 
 public struct TypeInfoOptions: OptionSet, Sendable { // add tests
-  public private(set) var rawValue: UInt32
+  public private(set) var rawValue: UInt16
   
-  public init(rawValue: UInt32) {
+  public init(rawValue: UInt16) {
     self.rawValue = rawValue
   }
